Перенесла команды по работе с заметками в Main
Настроила import
была ошибка: ANSWERS была описана после вызова main и это вызывало ошибку
