+++ notes_book.py
@@ -76,78 +76,6 @@ class NotesBook(UserList):
             result += note[0]+"\n" + note[1]+"\n"
         return result
 
-###################################################################
-# команды для handler - наверное надо перенести в основной модуль.
-# После переноса раскомментировать @error_handler
-# Я тестила: перенесла в основной модуль, добавила @error_handler,
-# и все функции NotesBook заработали
-###################################################################
-
-
-@error_handler
-def add_note():
-    print('Please input your note:')
-    # ввод многострочной заметки
-    lines = []
-    flag = True
-    while flag:
-        line = input()
-        if line:
-            lines.append(line)
-        else:
-            flag = False
-    text = '\n'.join(lines)
-    # ввод тєгов
-    hashtag = input('Please input the hashtag of your note: \n')
-    # добавление заметки в NotesBook
-    notes_book.add_note(text, hashtag.upper())
-    return "Your note is successfully saved"
-
-
-# @error_handler
-def delete_note():
-    print("Please input a hashtag of note that you would like to delete:")
-    hashtag = input().upper()
-    notes_book.delete_note(hashtag)
-    return f"The note with hashtag '{hashtag}' is deleted"
-
-
-# @error_handler
-def edit_note():
-    print("Please input a hashtag of note that you would like to edit:")
-    hashtag = input().upper()
-    notes_book.edit_note(hashtag)
-    return "The note is edited"
-
-
-# @error_handler
-def find_note():
-    print('Please input keyword for search:')
-    keyword = input().upper()
-    print('THE RESULTS OF SEARCH:')
-    print(notes_book.find_note(keyword))
-    return "The search is sucessfully finished"
-
-
-# @error_handler
-def sort_notes():
-    print("What type of sort would you like? Please input:")
-    print("1 - to sort from A to Z")
-    print("2 - to sort from Z to A")
-    print("3 - to sort from old notes to new notes")
-    print("4 - to sort from new notes to old notes")
-    search_type = input()
-    print('The sorted Notes are:')
-    print(notes_book.sort_notes(search_type))
-    return('The end of sorted Notes')
-
-
-# @error_handler
-def show_notes():
-    print('Your Notes Book:')
-    print(notes_book)
-    return "The end of Notes Book"
-
 
 def notes_main():
     # техническая функция для проверки работы модуля, можно удалить
